Fixed scrolls, tried to fix skillinfo... skillinfo still hangs my client...

+++ server/release/scripts/magic/scroll.py
@@ -7,8 +7,8 @@
 # Script for Spell Scrolls                                      #
 #===============================================================#
 
-from magic.spellbook import addspell
 import wolfpack
+from magic.spellbook import addspell, hasspell
 
 # calculate the spell-id of a scroll
 def calcSpellId( item ):
@@ -25,7 +25,7 @@ def onUse( char, item ):
 		char.message( 'Casting spell %d' % calcSpellId( item ) )
 	except:
 		char.socket.sysmessage( 'Broken scroll' )
-		return 0
+		return False
 
 	if item.amount > 1:
 		item.amount -= 1
@@ -33,31 +33,41 @@ def onUse( char, item ):
 	else:
 		item.delete()
 
-	return 1
+	return True
 
 def onDropOnItem( cont, item ):
 	# Who is dragging us?
 	char = item.container
 
 	if not char.ischar():
-		return 0
+		return False
+
+	if cont == char.getbackpack() or cont == item.container:
+		return False
 
 	try:
 		spell = calcSpellId( item )
 	except:
 		char.socket.sysmessage( "Error while calculating scroll id." )
-		return 0
+		return False
 
-	if spell >= 64:
-		char.socket.sysmessage( "Scroll with invalid spell-id: %d" % spell )
-		return 0
+	if ( 'magic.spellbook' in cont.events ):
+		if spell >= 64:
+			char.socket.sysmessage( "Scroll with invalid spell-id: %d" % spell )
+		else:
+			if not hasspell( cont, spell ):
+				addspell( cont, spell )
 
-	addspell( cont, spell )
+				if item.amount > 1:
+					item.amount -= 1
+					item.update()
+				else:
+					item.delete() # Consume the scroll
+				return False
+			else:
+				return False
 
-	if item.amount > 1:
-		item.amount -= 1
-		return 0
 	else:
-		item.delete() # Consume the scroll
+		return False
 
-	return 1
+	return False
